ENH:  harmonise step sizes in slider controls for Models module
Standardised the slider step sizes per 
https://discourse.slicer.org/t/segmentations-module-clicking-on-slider-track-should-move-slider-in-steps-spinner-should-increment-last-digit/20376/4?u=div
especially for opacity controls, but also for something to do with decimals.  

Note that some sliders only had singleStep explicitly defined, so that pageStep was implicitly defined.  
I think it would be a good practice to always explicitly define both singleStep and pageStep for each slider.  

QUERY:  do we need both 
        <property name="singleStep">
         <double>0.050000000000000</double>
        </property>
and
        <property name="SingleStep" stdset="0">
         <double>0.050000000000000</double>
        </property>
???
I would have thought that the latter supersedes the former.