Enhancement:  refine tooltips in Segment Editor module
Indicate that the user needs to be clicking 2D (planes/slices) when using the Islands tool, not in the 3D pane.  
The wording of the tooltips for "Remove selected island" and "Add selected island" now follows that of "Keep selected island".  
QUERY:  should it be possible for these actions to work in 3D???